MATLAB:mode:InvalidInput in how_to_use_wrapper
Fixed by turning matrix of uint32 integers to double. It seems to be
working but I’ll ask Alex if that changes anything for the MISC
algorithm.

+++ MISC Wrapper/how_to_use_wrapper.m
@@ -13,7 +13,7 @@ for curseq = 1:numseq
 end
 msa = aa2int(msa);
 
-msa = uint32(msa);
+msa = double(uint32(msa));
 %% get reference sequence
 refID = 'P11413';
 gensearch = getgenpept(refID);
